Move page listing view to wagtail.admin.views.pages.listing
